Rename category.py
